revert breaking change
