+++ src/services/pan.service.ts
@@ -8,7 +8,7 @@ export interface PanService {
   // json file.
   // tslint: disable-next-line:no-any
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  getDatasets(title: string): Promise<any>;
+  getDetails(title: string): Promise<any>;
 }
 
 export class PanProvider implements Provider<PanService> {
