+++ src/utils.ts
@@ -1,4 +1,6 @@
 import Qty = require('js-quantities');
+import {Filter, Where} from '@loopback/repository';
+import {Dataset} from './models';
 
 export interface Query {
   variable: string;
@@ -35,4 +37,74 @@ export function convertUnits(value: number, unit: string) {
 
 export function convertNameforScicat(panoscName: string) {
   return 'scientificMetadata.' + panoscName + '.value';
+}
+
+export function convertQueryForSciCat(filter?: Filter<Dataset>) {
+  const scicatQuery: Filter = {};
+  if (filter !== undefined && typeof filter !== undefined) {
+    if ('limit' in filter!) {
+      const limit = filter!['limit'];
+      if (limit !== undefined && typeof limit !== undefined) {
+        scicatQuery['limit'] = limit;
+      } else {
+        scicatQuery['limit'] = 1;
+      }
+    }
+    if ('skip' in filter!) {
+      const skip = filter!['skip'];
+      if (skip !== undefined && typeof skip !== undefined) {
+        scicatQuery['skip'] = skip;
+      } else {
+        scicatQuery['skip'] = 0;
+      }
+    }
+    const where = filter!.where;
+    if (where !== undefined && typeof where !== undefined) {
+      if ('and' in where) {
+        const parameterSearchArray: LoopBackQuery[] = [];
+        where.and.forEach((element: Object) => {
+          const query1 = element as Query;
+          console.log(query1);
+          const convertedValue = convertUnits(query1.value, query1.unit);
+          const convertedName = convertNameforScicat(query1.variable);
+          const andElement: Where = {
+            [convertedName]: {
+              [query1.operator]: convertedValue,
+            },
+          };
+          parameterSearchArray.push(andElement);
+        });
+        scicatQuery['where'] = {and: parameterSearchArray};
+      } else if ('or' in where) {
+        const parameterSearchArray: LoopBackQuery[] = [];
+        where.or.forEach((element: Object) => {
+          const query1 = element as Query;
+          console.log(query1);
+          const convertedValue = convertUnits(query1.value, query1.unit);
+          const andElement: Where = {
+            [query1.variable]: {
+              [query1.operator]: convertedValue,
+            },
+          };
+          parameterSearchArray.push(andElement);
+        });
+        scicatQuery['where'] = {or: parameterSearchArray};
+      } else if ('query' in where) {
+        const query2 = where!.query as Query;
+        const convertedValue = convertUnits(query2.value, query2.unit);
+        const condition: Where = {
+          [query2.variable]: {
+            [query2.operator]: convertedValue,
+          },
+        };
+        scicatQuery['where'] = condition;
+      } else {
+        // breakout
+      }
+    }
+  }
+  const jsonString = JSON.stringify(scicatQuery);
+  console.log(jsonString);
+  const jsonLimits = encodeURIComponent(jsonString);
+  return jsonLimits;
 }
